fix(webserver): rework tests in order to load only required flows
