+++ webserver/src/test/java/io/kestra/webserver/controllers/api/TaskRunControllerTest.java
@@ -1,6 +1,10 @@
 package io.kestra.webserver.controllers.api;
 
-import io.kestra.webserver.controllers.h2.JdbcH2ControllerTest;
+import static org.hamcrest.MatcherAssert.assertThat;
+import static org.hamcrest.Matchers.is;
+import static org.junit.jupiter.api.Assertions.assertThrows;
+
+import io.kestra.core.junit.annotations.KestraTest;
 import io.micronaut.http.HttpRequest;
 import io.micronaut.http.HttpStatus;
 import io.micronaut.http.client.annotation.Client;
@@ -9,11 +13,8 @@ import io.micronaut.reactor.http.client.ReactorHttpClient;
 import jakarta.inject.Inject;
 import org.junit.jupiter.api.Test;
 
-import static org.hamcrest.MatcherAssert.assertThat;
-import static org.hamcrest.Matchers.is;
-import static org.junit.jupiter.api.Assertions.*;
-
-class TaskRunControllerTest extends JdbcH2ControllerTest {
+@KestraTest
+class TaskRunControllerTest {
     @Inject
     @Client("/")
     private ReactorHttpClient client;
