--- webserver/src/test/java/io/kestra/webserver/controllers/h2/JdbcH2ControllerTest.java
@@ -1,41 +1,0 @@
-package io.kestra.webserver.controllers.h2;
-
-
-import io.kestra.core.repositories.LocalFlowRepositoryLoader;
-import io.kestra.core.runners.RunnerUtils;
-import io.kestra.core.runners.StandAloneRunner;
-import io.kestra.core.utils.TestsUtils;
-import io.kestra.jdbc.JdbcTestUtils;
-import io.kestra.core.junit.annotations.KestraTest;
-import jakarta.inject.Inject;
-import lombok.SneakyThrows;
-import org.junit.jupiter.api.BeforeEach;
-
-@KestraTest
-public class JdbcH2ControllerTest {
-    @Inject
-    protected StandAloneRunner runner;
-
-    @Inject
-    private JdbcTestUtils jdbcTestUtils;
-
-    @Inject
-    protected RunnerUtils runnerUtils;
-
-    @Inject
-    protected LocalFlowRepositoryLoader repositoryLoader;
-
-    @SneakyThrows
-    @BeforeEach
-    protected void setup() {
-        jdbcTestUtils.drop();
-        jdbcTestUtils.migrate();
-
-        TestsUtils.loads(repositoryLoader);
-
-        if (!runner.isRunning()) {
-            runner.setSchedulerEnabled(false);
-            runner.run();
-        }
-    }
-}
